fix training

+++ onmt/Trainer.py
@@ -45,10 +45,10 @@ class Statistics(object):
         return 100 * (self.n_correct / self.n_words)
 
     def mean_loss(self):
-        return self.loss / self.n_words
+        return self.loss / self.n_words.float()
 
     def ppl(self):
-        return math.exp(min(self.loss / self.n_words, 100))
+        return math.exp(min(self.loss / self.n_words.float(), 100))
 
     def elapsed_time(self):
         return time.time() - self.start_time
